+++ etc/env.dist
@@ -32,6 +32,8 @@ AWS_SECRET_ACCESS_KEY=changeme
 AWS_REGION=us-east-1
 TEMPLATE_BUCKET=changeme
 WEBSITE_BUCKET=changme
+SQS_CATEGORIZE_URL=
+SQS_CHECK_CATEGORY_URL=
 
 # COGNITO
 AWS_COGNITO_ENABLED=0
