--- src/utils/categorization/__init__.py
@@ -1,1 +1,0 @@
-"""Categorization Proxies."""
