--- src/utils/two_captcha.py
@@ -1,9 +1,0 @@
-"""Two Captcha utilies."""
-# Standard Python Libraries
-import os
-
-# Third-Party Libraries
-from twocaptcha import TwoCaptcha
-
-two_captcha_api_key = os.environ.get("TWO_CAPTCHA")
-solver = TwoCaptcha(two_captcha_api_key)
